+++ CHANGELOG.md
@@ -17,6 +17,8 @@ BUG FIXES:
   - providers/virtualbox: Retryable VBoxManage commands are properly
     retried. [GH-2365]
   - provisioners/ansible: Verbosity won't be blank by default. [GH-2320]
+  - provisioners/chef: Fix exception raised during Chef client node
+    cleanup. [GH-2345]
   - provisioners/salt: Correct master seed file name. [GH-2359]
 
 ## 1.3.4 (October 2, 2013)